requirements.txt

scikit-learn
scikit-surprise
pandas
numpy--1.26.4
